Bank the June R1/R2 render-submission perf work against 115 commits (~2.5
weeks) of parity + playtest progress on main. Single content conflict
(game/client/ClientWorld.cs); every other file auto-merged. Build clean, full
suite 2953/2953 green.

ClientWorld.cs — DriveEntityNodes + its call site:
- Kept cpuopt's R1/R2 central entity-node drive (DriveEntityNodes folds the
  entity PVS cull in and REPLACES the old ApplyEntityPvsCull) AND main's new
  DriveBodyAlpha (crosshair-chase local-body fade) call — both now run.
- Merged main's warpzone portal-exit cluster union into the per-node PVS
  visibility test, restructured so the local-player / camera-in-solid / cull-
  off case still falls through to DriveSync. main's version early-`continue`d
  there, which under the R1 central drive would have skipped those nodes'
  transform sync entirely (they'd never move).

EntityNode.cs — DriveSync dirty-gate (semantic fix, not a marker conflict):
- main added a time-driven MF_ROTATE key spin (Entity.ModelSpinRotate) inside
  SyncFromEntity, independent of ItemAnimate. cpuopt's dirty-gate only force-
  re-syncs while ItemAnimate != 0, so a stationary spinning pickup-key would
  have FROZEN. Added `!e.ModelSpinRotate` to the skip condition so it animates.

Next (unchanged by this merge): the R1/R2 entity-drive still hasn't had its
in-app visual playtest — do that before relying on it. Playtest-watch the two
other cpuopt gates now spanning main features: HudManager R8 hidden-panel
re-record + AmmoPanel R5 NeedsRedraw (glance at HUD under the scoreboard and
the ammo panel).

+++ game/client/particles/FaithfulParticleBackend.cs
@@ -172,7 +172,7 @@ public sealed partial class FaithfulParticleBackend : Node3D
     /// velmins==velmaxs). <paramref name="count"/> is DP's <c>pcount</c> (quality/countmultiplier applied
     /// inside the sim). Coordinates are Quake space (the sim's space).
     /// </summary>
-    public void Spawn(IReadOnlyList<EffectInfoEmitter> blocks, NVec3 origin, NVec3 velocity, int count,
+    public void Spawn(IReadOnlyList<EffectInfoEmitter> blocks, NVec3 origin, NVec3 velocity, float count,
         uint tintRgba = 0xFFFFFFFFu)
     {
         if (blocks is null || blocks.Count == 0)
@@ -187,7 +187,7 @@ public sealed partial class FaithfulParticleBackend : Node3D
     /// supplied <paramref name="velocity"/> is the emit velocity passed through velmins==velmaxs.
     /// <paramref name="count"/> is DP's <c>pcount</c>.
     /// </summary>
-    public void Trail(IReadOnlyList<EffectInfoEmitter> blocks, NVec3 start, NVec3 end, NVec3 velocity, int count,
+    public void Trail(IReadOnlyList<EffectInfoEmitter> blocks, NVec3 start, NVec3 end, NVec3 velocity, float count,
         uint tintRgba = 0xFFFFFFFFu)
     {
         if (blocks is null || blocks.Count == 0)
